🌱 refactor: split Workloads, GitOps, SyncDialog components (issue #21777)
Split four oversized components into X.tsx + X.parts.tsx + useX.ts
following the established RemediationConsole/DrillDownModal pattern.

- Workloads.tsx (570→126 lines): extract useWorkloads.ts + Workloads.parts.tsx + Workloads.types.ts
- GitOps.tsx (616→85 lines): extract useGitOps.ts + GitOps.parts.tsx + GitOps.types.ts
- SyncDialog.tsx (495→124 lines): extract useSyncDialog.ts + SyncDialog.parts.tsx
