+++ web/src/components/gitops/SyncDialog.tsx
@@ -1,75 +1,12 @@
-import { useState, useEffect, useRef, startTransition } from 'react'
-import { useTranslation } from 'react-i18next'
-import { Check, AlertTriangle, Play, Loader2, ChevronRight, GitBranch, Box, Server, Shield, Settings, Database, Network, Layers, Container, FileText, Puzzle, X } from 'lucide-react'
+import { Check, Play, GitBranch } from 'lucide-react'
 import { BaseModal } from '../../lib/modals'
-import { TechnicalAcronym } from '../shared/TechnicalAcronym'
-import { FETCH_DEFAULT_TIMEOUT_MS, LOCAL_AGENT_HTTP_URL } from '../../lib/constants'
-import { getStoredAuthToken } from '../../lib/authToken'
-import { agentFetch } from '../../hooks/mcp/shared'
-
-// Sync phases
-type SyncPhase = 'detection' | 'plan' | 'execution' | 'complete'
-
-// Get icon for Kubernetes resource kind
-function getResourceIcon(kind: string) {
-  const k = kind?.toLowerCase() || ''
-  if (k.includes('deployment')) return <Box className="w-4 h-4 text-blue-400" />
-  if (k.includes('service')) return <Network className="w-4 h-4 text-green-400" />
-  if (k.includes('pod')) return <Container className="w-4 h-4 text-cyan-400" />
-  if (k.includes('configmap')) return <Settings className="w-4 h-4 text-purple-400" />
-  if (k.includes('secret')) return <Shield className="w-4 h-4 text-red-400" />
-  if (k.includes('serviceaccount')) return <Server className="w-4 h-4 text-orange-400" />
-  if (k.includes('role') || k.includes('clusterrole')) return <Shield className="w-4 h-4 text-yellow-400" />
-  if (k.includes('customresourcedefinition') || k.includes('crd')) return <Puzzle className="w-4 h-4 text-purple-400" />
-  if (k.includes('namespace')) return <Layers className="w-4 h-4 text-blue-400" />
-  if (k.includes('persistentvolume') || k.includes('pvc')) return <Database className="w-4 h-4 text-cyan-400" />
-  if (k.includes('ingress')) return <Network className="w-4 h-4 text-green-400" />
-  if (k.includes('statefulset') || k.includes('daemonset') || k.includes('replicaset')) return <Layers className="w-4 h-4 text-blue-400" />
-  if (k.includes('job') || k.includes('cronjob')) return <Settings className="w-4 h-4 text-yellow-400" />
-  if (k.includes('webhook')) return <Network className="w-4 h-4 text-purple-400" />
-  return <FileText className="w-4 h-4 text-yellow-500" />
-}
-
-// Format resource kind for display
-function formatResourceKind(kind: string): string {
-  if (!kind) return 'Resource'
-  // Common abbreviations - these will be wrapped with tooltips where they're rendered
-  if (kind.toLowerCase() === 'customresourcedefinition') return 'CRD'
-  if (kind.toLowerCase() === 'serviceaccount') return 'ServiceAccount'
-  if (kind.toLowerCase() === 'clusterrole') return 'ClusterRole'
-  if (kind.toLowerCase() === 'clusterrolebinding') return 'ClusterRoleBinding'
-  return kind
-}
-
-// Wrapper to format resource kind with tooltip if it's an acronym
-function FormattedResourceKind({ kind }: { kind: string }) {
-  const formatted = formatResourceKind(kind)
-  if (formatted === 'CRD') {
-    return <TechnicalAcronym term="CRD">CRD</TechnicalAcronym>
-  }
-  return <>{formatted}</>
-}
-
-interface DriftedResource {
-  kind: string
-  name: string
-  namespace: string
-  field: string
-  gitValue: string
-  clusterValue: string
-}
-
-interface SyncPlan {
-  action: 'create' | 'update' | 'delete'
-  resource: string
-  details: string
-}
-
-interface SyncLogEntry {
-  timestamp: string
-  message: string
-  status: 'pending' | 'running' | 'success' | 'error'
-}
+import { useSyncDialog } from './useSyncDialog'
+import {
+  SyncPhaseIndicator,
+  DetectionPhaseContent,
+  PlanPhaseContent,
+  ExecutionPhaseContent,
+} from './SyncDialog.parts'
 
 interface SyncDialogProps {
   isOpen: boolean
@@ -90,221 +27,21 @@ export function SyncDialog({
   cluster,
   repoUrl,
   path,
-  onSyncComplete }: SyncDialogProps) {
-  const { t } = useTranslation()
-  const [phase, setPhase] = useState<SyncPhase>('detection')
-  const [driftedResources, setDriftedResources] = useState<DriftedResource[]>([])
-  const [syncPlan, setSyncPlan] = useState<SyncPlan[]>([])
-  const [syncLogs, setSyncLogs] = useState<SyncLogEntry[]>([])
-  const [tokenCount, setTokenCount] = useState(0)
-  const [error, setError] = useState<string | null>(null)
-  const [isInitializing, setIsInitializing] = useState(false)
-  const logContainerRef = useRef<HTMLDivElement>(null)
-
-  // Auto-scroll logs
-  useEffect(() => {
-    if (logContainerRef.current) {
-      logContainerRef.current.scrollTop = logContainerRef.current.scrollHeight
-    }
-  }, [syncLogs])
-
-  // Note: ESC key handling is now handled by BaseModal
-
-  const addLog = (message: string, status: SyncLogEntry['status'] = 'pending') => {
-    const entry: SyncLogEntry = {
-      timestamp: new Date().toLocaleTimeString(),
-      message,
-      status }
-    setSyncLogs(prev => [...prev, entry])
-  }
-
-  const updateLastLog = (status: SyncLogEntry['status']) => {
-    setSyncLogs(prev => {
-      if (prev.length === 0) return prev
-      const updated = [...prev]
-      updated[updated.length - 1] = { ...updated[updated.length - 1], status }
-      return updated
-    })
-  }
-
-  // Phase 1: Detection
-  const runDetection = async () => {
-    setIsInitializing(true)
-    addLog('Connecting to cluster...', 'running')
-
-    try {
-      // #7993 Phase 4: detect-drift moved to kc-agent. Runs under the
-      // user's kubeconfig instead of the backend pod SA.
-      const token = await getStoredAuthToken()
-      const response = await agentFetch(`${LOCAL_AGENT_HTTP_URL}/gitops/detect-drift`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          'X-Requested-With': 'XMLHttpRequest',
-          ...(token && { 'Authorization': `Bearer ${token}` }) },
-        body: JSON.stringify({
-          repoUrl,
-          path,
-          cluster,
-          namespace }),
-        signal: AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS) })
-
-      updateLastLog('success')
-      addLog('Analyzing drift...', 'running')
-
-      if (!response.ok) {
-        const err = await response.json()
-        throw new Error(err.error || 'Failed to detect drift')
-      }
-
-      const data = await response.json()
-      updateLastLog('success')
-
-      // Parse the response
-      if (data.resources && data.resources.length > 0) {
-        setDriftedResources(data.resources)
-        addLog(`Found ${data.resources.length} drifted resources`, 'success')
-      } else if (data.drifted) {
-        // Fallback: create a generic drift entry from raw diff
-        const genericDrift: DriftedResource[] = [{
-          kind: 'Resource',
-          name: appName,
-          namespace,
-          field: 'configuration',
-          gitValue: 'git state',
-          clusterValue: 'cluster state' }]
-        setDriftedResources(genericDrift)
-        addLog('Drift detected (see raw diff)', 'success')
-      } else {
-        addLog('No drift detected - cluster is in sync', 'success')
-        setDriftedResources([])
-      }
-
-      if (data.tokensUsed) {
-        setTokenCount(prev => prev + data.tokensUsed)
-      }
-
-      setPhase('plan')
-    } catch (err: unknown) {
-      // Error states should be shown immediately (not deferred with startTransition)
-      updateLastLog('error')
-      const message = err instanceof Error ? err.message : 'Detection failed'
-      addLog(`Error: ${message}`, 'error')
-      setError(message)
-    } finally {
-      // Always reset initializing state
-      setIsInitializing(false)
-    }
-  }
-
-  // Reset state when dialog opens
-  useEffect(() => {
-    if (isOpen) {
-      // Clear error immediately so user doesn't see stale errors
-      setError(null)
-      // Batch non-critical state resets to prevent flicker
-      startTransition(() => {
-        setPhase('detection')
-        setDriftedResources([])
-        setSyncPlan([])
-        setSyncLogs([])
-        setTokenCount(0)
-      })
-      // runDetection() will set isInitializing to true
-      runDetection()
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isOpen])
-
-  // Phase 2: Generate Plan
-  useEffect(() => {
-    if (phase === 'plan' && driftedResources.length > 0) {
-      const plan: SyncPlan[] = driftedResources.map(r => ({
-        action: 'update' as const,
-        resource: `${r.kind}/${r.name}`,
-        details: `${r.field}: ${r.clusterValue} → ${r.gitValue}` }))
-      setSyncPlan(plan)
-    }
-  }, [phase, driftedResources])
-
-  // Phase 3: Execute Sync
-  const runSync = async () => {
-    setPhase('execution')
-    addLog('Starting sync...', 'running')
-
-    try {
-      // #7993 Phase 4: gitops sync moved to kc-agent. Runs under the
-      // user's kubeconfig instead of the backend pod SA.
-      const token = await getStoredAuthToken()
-      const response = await agentFetch(`${LOCAL_AGENT_HTTP_URL}/gitops/sync`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          'X-Requested-With': 'XMLHttpRequest',
-          ...(token && { 'Authorization': `Bearer ${token}` }) },
-        body: JSON.stringify({
-          repoUrl,
-          path,
-          cluster,
-          namespace }),
-        signal: AbortSignal.timeout(FETCH_DEFAULT_TIMEOUT_MS) })
-
-      updateLastLog('success')
-
-      if (!response.ok) {
-        const err = await response.json()
-        throw new Error(err.error || 'Sync failed')
-      }
-
-      const data = await response.json()
-
-      // Log applied resources
-      if (data.applied && data.applied.length > 0) {
-        for (const resource of data.applied) {
-          addLog(`✓ ${resource}`, 'success')
-        }
-      }
-
-      // Log any errors
-      if (data.errors && data.errors.length > 0) {
-        for (const error of data.errors) {
-          addLog(`✗ ${error}`, 'error')
-        }
-      }
-
-      if (data.tokensUsed) {
-        setTokenCount(prev => prev + data.tokensUsed)
-      }
-
-      if (data.success) {
-        addLog('Sync complete!', 'success')
-        setPhase('complete')
-      } else {
-        addLog(`Sync failed: ${data.message}`, 'error')
-        setError(data.message)
-      }
-    } catch (err: unknown) {
-      updateLastLog('error')
-      const message = err instanceof Error ? err.message : 'Sync failed'
-      addLog(`Error: ${message}`, 'error')
-      setError(message)
-    }
-  }
-
-  const handleClose = () => {
-    if (phase === 'complete') {
-      onSyncComplete()
-    }
-    onClose()
-  }
-
-  const phaseProgress = {
-    detection: 1,
-    plan: 2,
-    execution: 3,
-    complete: 4 }
-
-  const isSyncing = phase === 'plan' || phase === 'execution'
+  onSyncComplete,
+}: SyncDialogProps) {
+  const {
+    phase,
+    driftedResources,
+    syncPlan,
+    syncLogs,
+    tokenCount,
+    error,
+    isInitializing,
+    logContainerRef,
+    isSyncing,
+    runSync,
+    handleClose,
+  } = useSyncDialog({ isOpen, appName, namespace, cluster, repoUrl, path, onSyncComplete, onClose })
 
   return (
     <BaseModal isOpen={isOpen} onClose={handleClose} size="lg" closeOnBackdrop={!isSyncing} closeOnEscape={!isSyncing}>
@@ -316,139 +53,31 @@ export function SyncDialog({
         showBack={false}
       />
 
-      {/* Phase Indicator */}
-      <div className="px-6 py-3 bg-muted/30 border-b border-border">
-          <div className="flex items-center justify-between text-sm">
-            {['Detection', 'Plan', 'Execute', 'Complete'].map((label, i) => {
-              const stepNum = i + 1
-              const isActive = phaseProgress[phase] === stepNum
-              const isComplete = phaseProgress[phase] > stepNum
-
-              return (
-                <div key={label} className="flex items-center gap-2">
-                  <div className={`w-6 h-6 rounded-full flex items-center justify-center text-xs font-medium
-                    ${isComplete ? 'bg-green-500 text-foreground' :
-                      isActive ? 'bg-primary text-primary-foreground' :
-                      'bg-muted text-muted-foreground'}`}
-                  >
-                    {isComplete ? <Check className="w-3 h-3" /> : stepNum}
-                  </div>
-                  <span className={isActive ? 'text-foreground font-medium' : 'text-muted-foreground'}>
-                    {label}
-                  </span>
-                  {i < 3 && <ChevronRight className="w-4 h-4 text-muted-foreground mx-2" />}
-                </div>
-              )
-            })}
-          </div>
-        </div>
+      <SyncPhaseIndicator phase={phase} />
 
       <BaseModal.Content className="max-h-[400px]">
-          {/* Initial Loading State */}
-          {isInitializing && phase === 'detection' && syncLogs.length === 0 && (
-            <div className="flex flex-col items-center justify-center py-8 space-y-3">
-              <Loader2 className="w-8 h-8 animate-spin text-blue-400" />
-              <p className="text-sm text-muted-foreground">Initializing drift detection...</p>
-            </div>
-          )}
+        {phase === 'detection' && (
+          <DetectionPhaseContent
+            isInitializing={isInitializing}
+            syncLogsLength={syncLogs.length}
+          />
+        )}
 
-          {/* Detection Phase */}
-          {phase === 'detection' && syncLogs.length > 0 && (
-            <div className="space-y-3">
-              <div className="flex items-center gap-2 text-muted-foreground">
-                <Loader2 className="w-4 h-4 animate-spin" />
-                <span>{t('gitops.detectingDrift')}</span>
-              </div>
-            </div>
-          )}
+        {phase === 'plan' && (
+          <PlanPhaseContent
+            driftedResources={driftedResources}
+            syncPlan={syncPlan}
+          />
+        )}
 
-          {/* Plan Phase */}
-          {phase === 'plan' && (
-            <div className="space-y-4">
-              <div>
-                <h3 className="text-sm font-medium text-foreground mb-3 flex items-center gap-2">
-                  <AlertTriangle className="w-4 h-4 text-yellow-500" />
-                  Drift Detected ({driftedResources.length} resources)
-                </h3>
-                <div className="space-y-2 max-h-[250px] overflow-y-auto">
-                  {driftedResources.map((r, i) => (
-                    <div key={i} className="p-3 rounded-lg bg-yellow-500/10 border border-yellow-500/20">
-                      <div className="flex items-center gap-2 text-sm">
-                        {getResourceIcon(r.kind)}
-                        <span className="px-1.5 py-0.5 rounded text-xs font-medium bg-card text-muted-foreground">
-                          <FormattedResourceKind kind={r.kind} />
-                        </span>
-                        <span className="font-medium text-foreground truncate">{r.name}</span>
-                      </div>
-                      {(r.field || r.clusterValue || r.gitValue) && (
-                        <div className="mt-2 text-xs font-mono pl-6">
-                          {r.field && <span className="text-muted-foreground">{r.field}: </span>}
-                          {r.clusterValue && <span className="text-red-400 line-through">{r.clusterValue}</span>}
-                          {r.clusterValue && r.gitValue && <span className="text-muted-foreground"> → </span>}
-                          {r.gitValue && <span className="text-green-400">{r.gitValue}</span>}
-                        </div>
-                      )}
-                    </div>
-                  ))}
-                </div>
-              </div>
+        {(phase === 'execution' || phase === 'complete') && (
+          <ExecutionPhaseContent
+            syncLogs={syncLogs}
+            tokenCount={tokenCount}
+            logContainerRef={logContainerRef}
+          />
+        )}
 
-              <div>
-                <h3 className="text-sm font-medium text-foreground mb-3">Sync Plan ({syncPlan.length} changes)</h3>
-                <div className="space-y-1 max-h-[120px] overflow-y-auto">
-                  {syncPlan.map((item, i) => {
-                    const [kind, name] = item.resource.includes('/') ? item.resource.split('/') : ['Resource', item.resource]
-                    return (
-                      <div key={i} className="flex items-center gap-2 text-sm py-1">
-                        <span className={`px-1.5 py-0.5 rounded text-xs font-medium ${
-                          item.action === 'create' ? 'bg-green-500/20 text-green-400' :
-                          item.action === 'delete' ? 'bg-red-500/20 text-red-400' :
-                          'bg-blue-500/20 text-blue-400'
-                        }`}>
-                          {item.action.toUpperCase()}
-                        </span>
-                        {getResourceIcon(kind)}
-                        <span className="text-muted-foreground text-xs"><FormattedResourceKind kind={kind} /></span>
-                        <span className="text-foreground truncate">{name}</span>
-                      </div>
-                    )
-                  })}
-                </div>
-              </div>
-            </div>
-          )}
-
-          {/* Execution Phase */}
-          {(phase === 'execution' || phase === 'complete') && (
-            <div className="space-y-4">
-              <div className="flex items-center justify-between text-sm">
-                <span className="text-muted-foreground">Console Output</span>
-                <span className="text-xs px-2 py-1 rounded bg-muted text-muted-foreground">
-                  Tokens: {tokenCount.toLocaleString()}
-                </span>
-              </div>
-              <div
-                ref={logContainerRef}
-                className="h-48 p-3 rounded-lg bg-black/50 border border-border font-mono text-xs overflow-y-auto"
-              >
-                {syncLogs.map((log, i) => (
-                  <div key={i} className="flex items-start gap-2 py-0.5">
-                    <span className="text-muted-foreground shrink-0">{log.timestamp}</span>
-                    {log.status === 'running' && <Loader2 className="w-3 h-3 animate-spin text-blue-400 mt-0.5" />}
-                    {log.status === 'success' && <Check className="w-3 h-3 text-green-400 mt-0.5" />}
-                    {log.status === 'error' && <X className="w-3 h-3 text-red-400 mt-0.5" />}
-                    <span className={
-                      log.status === 'success' ? 'text-green-400' :
-                      log.status === 'error' ? 'text-red-400' :
-                      'text-foreground'
-                    }>{log.message}</span>
-                  </div>
-                ))}
-              </div>
-            </div>
-          )}
-
-        {/* Error State */}
         {error && (
           <div className="p-4 rounded-lg bg-red-500/10 border border-red-500/20 text-red-400">
             {error}
